% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line.R
\name{concatenate}
\alias{concatenate}
\title{Concatenate Two Graph Embeddings}
\usage{
concatenate(input_one, input_two, binary = 0)
}
\arguments{
\item{input_one}{the first numeric matrix returned by line. This should have a 
different order than input_two}

\item{input_two}{the second numeric matrix returned by line. This should have a 
different order than input_one}

\item{binary}{This should always be zero as we don't want a binary formatted dataframe. default is#'  0}
}
\description{
This function concatenates two graph embeddings represented as numeric matrixes returned by 
line
}
\details{
This function concatenates two graph embeddings represented as numeric matrixes of differnet
orders. It takes two numeric matrixes returned by line after calling line with the same input
dataframe but different orders (1 and 2). It returns a combined graph embedding represented 
as a numeric matrix with row names representing the vertice. Each row is a 
weight of the row name/feature/vertice.
}
\examples{
u <- c("good", "the", "bad")
v <- c("the", "good", "the")
w <- 1:3
df <- data.frame(u, v, w)
new_df <- reconstruct(df)
order_1 <- line(df = new_df, binary = 0, dim = 100, order = 1, negative = 5, samples = 10, rho = 0.025, threads = 1)
order_2 <- line(df = new_df, binary = 0, dim = 100, order = 2, negative = 5, samples = 10, rho = 0.025, threads = 1) 
concatenate_df <- concatenate(input_one = order_1, input_two = order_2, binary = 0)
}
\references{
\url{https://arxiv.org/abs/1503.03578}
}
\seealso{
\url{https://github.com/tangjianpku/LINE}
}
