% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line.R
\name{normalize}
\alias{normalize}
\title{Normalize Graph Embedding}
\usage{
normalize(input_matrix)
}
\arguments{
\item{input_matrix}{numeric matrix that will be normalized.}
}
\description{
This function normalizes each row vector with the formula row = row / || row ||.
}
\details{
This function normalizes each row vector with the formula row = row / || row ||. This website
\url{https://www.khanacademy.org/computing/computer-programming/programming-natural-simulations/programming-vectors/a/vector-magnitude-normalization} is a good explanation of this normalization.
The output has floating point differences with the original line algorithm's normalize function
because R has different floating point arithmetic than C++.
}
\examples{
u <- c("good", "the", "bad")
v <- c("the", "good", "the")
w <- 1:3
df <- data.frame(u, v, w)
new_df <- reconstruct(df)
order_1 <- line(df = new_df, binary = 0, dim = 100,
          order = 1, negative = 5, samples = 1, rho = 0.025, threads = 1)
order_2 <- line(df = new_df, binary = 0, dim = 100,
          order = 2, negative = 5, samples = 1, rho = 0.025, threads = 1) 
concatenate_matrix <- concatenate(input_one = order_1, 
                     input_two = order_2, binary = 0)
normalize_matrix <- normalize(input_matrix = concatenate_matrix)
}
\references{
\url{https://arxiv.org/abs/1503.03578}
}
\seealso{
\url{https://github.com/tangjianpku/LINE}
}
